Closes #25.
Resolves issue with tuning RTL2832 based tuners into the range of
799.555894 and 801.584395.  The pre-defined sample rates were slightly
off and causing issues with this tuning range.  I recalculated the 
sample rates and removed any that were not integral multiples of the 
48,000 Hz desired channel rate.

Updated the sample bit depth for RTL2832 and HackRF tuners.  Even though
the sample bit depth is 8 bits for each, the displayed noise floor was
off, so I increased the bit depth for each.