Adds Remez filter designer for low, high and band-pass filters.  Support for Hilbert and Differentiator FIR filter designers may be added in the future.
Adds developer filter viewer tool written in JavaFX to visualize the
frequency and impulse response of FIR filters.  This application is not
intended to be a user application and is specifically for the developer
to analyze filters and filter specifications that will be used in the
code.

NOTE: Java 8 JDK is now required to build this project for version 0.3.x
and forward.  The JavaFX visualizer application and future user
interfaces will use the Java 8 version of JavaFX