WIP - Updated P25 C4FM and LSM decoders to assemble dibits and push reusable byte buffers out to the processing chain.  This enables the processing chain to broadcast these buffers back to the decoder for message assembly and to any running binary recorder.
Updated P25Decoder to be a reusable byte buffer listener.  The decoder
will now pass reusable byte buffers down to the message framer for
message processing.  This should make the code more efficient since it
can process byte arrays versus handling each individual dibit.

The P25MessageFramer has been updated to be a listener of reusable byte
buffers instead of Dibits, but the actual processing of byte arrays
has to be coded.

Added max recording size to the binary recorder so that it cycles at
500 kB max file size.