581 Add Support for Harris Vendor-Specific TSBK Messages (#582)
* Adds support for Harris Phase II systems.

Disables auto-detection of OSP/ISP from status symbols in the channel
status processor.  This is unreliable on Harris systems.

Updates phase 1 decoder state to not allocate traffic channels if the
channel is not backed by a frequency band message.

Disables auto-detection of OSP/ISP from status symbols in the channel
status processor.  This is unreliable on Harris systems.

Updates phase 1 decoder state to not allocate traffic channels if the
channel is not backed by a frequency band message.