#1118 LRRP Latitude Hemisphere parsing.  Updates parser to treat latitude value as a 1-bit hemisphere flag followed by a 31-bit integer.  Previously, this field was parsed as a 32-bit two's complement field.
There may be two separate formats that are used for LRRP position fields where one treats the value as a two's complement field and the other as a 1-bit hemisphere indicator followed by a 31-bit positive integer.  If so, we'll need to expose a user-selectable flag to toggle one format or the other.

For now, this change won't impact any Northern hemisphere users.